`Refactor: Update User and Product DTOs and Model to use Double instead of BigDecimal for price field`

+++ src/main/java/zone01/soufian/lets_play/dto/product/ProductResponse.java
@@ -1,10 +1,8 @@
 package zone01.soufian.lets_play.dto.product;
 
-import java.math.BigDecimal;
-
 public record ProductResponse(
     String id,
     String name,
     String description,
-    BigDecimal price
+    Double price
 ) {}
